refactor home view and create home view model
