+++ FitnessApp/Home/Model/Activity.swift
@@ -1,0 +1,18 @@
+//
+//  Activity.swift
+//  FitnessApp
+//
+//  Created by Balogun Kayode on 06/11/2024.
+//
+
+import SwiftUI
+
+struct Activity {
+    let id: Int
+    let title: String
+    let subtitle: String
+    let image: String
+    let tintColor: Color
+    let amount: String
+
+}
